refactor(#160): Repository 추상화
