+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/repository/login/TokenRepositoryImpl.kt
@@ -1,0 +1,66 @@
+package boostcamp.and07.mindsync.data.repository.login
+
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.core.edit
+import androidx.datastore.preferences.core.emptyPreferences
+import androidx.datastore.preferences.core.stringPreferencesKey
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.map
+import java.io.IOException
+import javax.inject.Inject
+
+class TokenRepositoryImpl
+@Inject
+constructor(
+    private val dataStore: DataStore<Preferences>,
+) : TokenRepository {
+    companion object {
+        private val ACCESS_TOKEN_KEY = stringPreferencesKey(DataStoreConst.ACCESS_TOKEN_KEY_NAME)
+        private val REFRESH_TOKEN_KEY = stringPreferencesKey(DataStoreConst.REFRESH_TOKEN_KEY_NAME)
+    }
+
+    override fun getAccessToken(): Flow<String?> {
+        return dataStore.data.catch { exception ->
+            if (exception is IOException) {
+                exception.printStackTrace()
+                emit(emptyPreferences())
+            } else {
+                throw exception
+            }
+        }.map { prefs ->
+            prefs[ACCESS_TOKEN_KEY]
+        }
+    }
+
+    override fun getRefreshToken(): Flow<String?> {
+        return dataStore.data.map { prefs ->
+            prefs[REFRESH_TOKEN_KEY]
+        }
+    }
+
+    override suspend fun saveAccessToken(token: String) {
+        dataStore.edit { prefs ->
+            prefs[ACCESS_TOKEN_KEY] = token
+        }
+    }
+
+    override suspend fun saveRefreshToken(token: String) {
+        dataStore.edit { prefs ->
+            prefs[REFRESH_TOKEN_KEY] = token
+        }
+    }
+
+    override suspend fun deleteAccessToken() {
+        dataStore.edit { prefs ->
+            prefs.remove(ACCESS_TOKEN_KEY)
+        }
+    }
+
+    override suspend fun deleteRefreshToken() {
+        dataStore.edit { prefs ->
+            prefs.remove(REFRESH_TOKEN_KEY)
+        }
+    }
+}
